+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Optional command line parameter `-ld` for specifying source library directory.
 
+### Changed
+
+- When literal label is located at the beginning of the line, but not followed by equal sign then it is considered a
+  label usage instead of definition and processing doesn't throw an error.
+
 ## [1.0.0] - 2024-12-12
 
 Initial public release.